chore: assert no exceptions logged during test run (#311)
The provider is intentionally resilient to malformed records, and
implements this resilience by logging and continuing on when any
exception is thrown. However, this prevents the unit tests from many
errors in the implementation of the provider. Therefore, spy on the
provider and fail unit tests if it is logging exceptions.

Signed-off-by: Will Murphy <will.murphy@anchore.com>
Signed-off-by: Andrew Nesbitt <andrewnez@gmail.com>